requirements.txt
# local package
-e .

# project requirements
numpy==1.23.4
pandas==1.5.1
python-math==0.0.1
tqdm==4.64.1
opencv-python==4.6.0.66
torch==1.13.0
torchvision==0.14.0
torchmetrics==0.10.2
torchsummary==1.5.1
matplotlib==3.6.2
seaborn==0.12.1
joblib==1.2.0
pytorch-metric-learning==1.6.3
pytorch-lightning==1.8.3.post1
wandb==0.13.5
umap-learn==0.5.3
pretty-confusion-matrix==0.1.1